Added visual improvements and user defined difficulty
Implemented a large variety of visual improvements to the worker including;
 - logo/header ASCII art
 - terminal output colors Green=solution, Yellow=alert, Red=error
 - added Worker Status
 - solution output (excl. Haiku) reduced to one line
 - separation of Host's required diff and solution diff, rdiff | sdiff
 - stabilized haikurate by averaging with the last calculation
 - changed debug data to only be visible on Trace=1
 - removed extraneous verbose option
 - output formatting fixes
Implemented auto and manual (user defined difficulty). If manual
difficulty is below the Host's required difficulty, the difficulty will
be switched back to auto.
Polling is speed up tenfold when there is no work so the worker is not
idle for a period of time when new work becomes available.
Fixed random seed (re)intialization when new work is received.
Rewrote get/send work protocol (hopefully increasing clarity).